+++ scripts/__init__.py
@@ -1,0 +1,3 @@
+"""
+This module contains all the scripts for the OpenFoodFacts data analysis project.
+"""
