+++ scripts/models/__init__.py
@@ -1,0 +1,3 @@
+"""
+This module handles data loading and basic data operations.
+"""
